Getting ready for another release.
During the boot stage the filsystem type was used to determine whether
to use the genext2fs (non-root) method or the normal method used by root.
Now the user type is used, this is better because it prevents conflicts
if a different filesystem type is chosen, or if a non-root user changes to a
different filesystem type in the filesystem box and forgets to change back
to genext2fs; however, it prevents root from using the genext2fs method,
which isn't a problem.  The sudo requirement for lilo is too much of a
requirement, and will be changed in future versions.